(10) - Emulator 1 - Type 1 (Almost done)

+++ src/debs/dataset.py
@@ -22,6 +22,7 @@ import xarray
 import numpy as np
 from   tools import get_data_path, get_mesh_path
 
+
 class BlackSea_Dataset():
     r"""A simple tool to load data of Black Sea simulations (NEMO Simulator)."""
 
@@ -155,8 +156,8 @@ class BlackSea_Dataset():
         # Creation of the mesh
         x_mesh, y_mesh = np.meshgrid(np.linspace(0, 1, num = x), np.linspace(0, 1, num = y), indexing = 'ij')
 
-        # Concatenation of the mesh
-        return np.stack((x_mesh, y_mesh), axis = 0)
+        # Concatenation of the mesh (np.float32 is the type needed for torch when converted afterforwards)
+        return np.stack((x_mesh, y_mesh), axis = 0, dtype = np.float32)
 
     def get_depth(self):
         r"""Used to retreive the bathymetry information, i.e. the depth in [m] for each region of the sea (2D)"""
@@ -165,7 +166,7 @@ class BlackSea_Dataset():
         path_mesh = get_mesh_path()
 
         # Loading the dataset containing information about the Black Sea mesh
-        return xarray.open_dataset(path_mesh, engine = "h5netcdf").bathy_metry.data
+        return xarray.open_dataset(path_mesh, engine = "h5netcdf").bathy_metry.data.astype('float32')
 
     def get_bathymetry(self):
         r"""Used to retreive the bathymetry mask, i.e. the depth index at which we reach the bottom of the sea (2D)"""
